chore: refactor packages
