+++ cmd/root.go
@@ -8,10 +8,16 @@ import (
 )
 
 // options
-var Configfile string
+
+var (
+	id         string
+	nowait     bool
+	pollsec    int
+	configfile string
+)
 
 // rootCmd represents the base command when called without any subcommands
-var RootCmd = &cobra.Command{
+var rootCmd = &cobra.Command{
 	Use:   "codebuild-multirunner",
 	Short: "This is a simple CLI tool to \"Start build with overrides\" multiple AWS CodeBuild Projects at once.",
 	Long: `This is a simple CLI tool to "Start build with overrides" multiple AWS CodeBuild Projects at once.
@@ -21,17 +27,17 @@ This command will read YAML based config file and run multiple CodeBuild project
 }
 
 func Execute() {
-	err := RootCmd.Execute()
+	err := rootCmd.Execute()
 	if err != nil {
 		os.Exit(1)
 	}
 }
 
 func init() {
-	RootCmd.PersistentFlags().StringVar(&Configfile, "config", "./.codebuild-multirunner.yaml", "file path for config file.")
+	rootCmd.PersistentFlags().StringVar(&configfile, "config", "./.codebuild-multirunner.yaml", "file path for config file.")
 }
 
 // set version from goreleaser variables
 func SetVersionInfo(version, commit, date string) {
-	RootCmd.Version = fmt.Sprintf("%s (Built on %s from Git SHA %s)", version, date, commit)
+	rootCmd.Version = fmt.Sprintf("%s (Built on %s from Git SHA %s)", version, date, commit)
 }
